web-app/server/requirements.txt
absl-py
astunparse
blinker
bcrypt
certifi
charset-normalizer
click
contourpy
cycler
Flask
flask-cors
flatbuffers
fonttools
gast
google-pasta
grpcio
h5py
idna
importlib_resources
itsdangerous
Jinja2
joblib
keras
kiwisolver
libclang
lxml
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml_dtypes
namex
nibabel
nilearn
numpy
opencv-contrib-python
opencv-python
opt_einsum
optree
packaging
pandas
pillow
protobuf
Pygments
pyparsing
python-dateutil
pytz
requests
rich
scikit-learn
scipy
seaborn
six
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
typing_extensions
tzdata
urllib3
Werkzeug
wrapt
